debug: double 1.0 to float 1.0 (activation\Sigmoid.c)

+++ activations/sigmoid.c
@@ -5,11 +5,11 @@
 void sigmoidForward(float *x, float *y, int dataLen)
 {
     for (int i=0; i<dataLen; i++)
-        *(y + i) = 1.0 / (1.0 + expf(-*(x + i)));
+        *(y + i) = (float)1.0 / ((float)1.0 + expf(-*(x + i)));
 }
 
 void sigmoidBackward(float *deltaLoss, float *Loss, float *y, int dataLen)
 {
     for (int i=0; i<dataLen; i++)
-        *(deltaLoss + i) = *(Loss + i) * (1.0 - *(y + i)) * *(y + i);
+        *(deltaLoss + i) = *(Loss + i) * ((float)1.0 - *(y + i)) * *(y + i);
 }
